pattern cancelling

+++ clip/cancel_jobs.py
@@ -13,14 +13,20 @@ def main():
                              action="store_true" )
     argparser.add_argument ( '-s','--server', help='cancel all server jobs',
                              action="store_true" )
+    argparser.add_argument ( '-d','--dry_run', help='dry-run just tell what you would do',
+                             action="store_true" )
     argparser.add_argument ( '-r','--run', help='cancel all RUN jobs (ie the walkers only)',
                              action="store_true" )
     argparser.add_argument ( '-H','--hiscore', help='cancel all hiscore jobs',
                              action="store_true" )
+    argparser.add_argument ( '-P', '--pattern', help='cancel all jobs that have <pattern> in the name',
+                             type=str, default=None )
     args=argparser.parse_args()
     grp = "| grep QOSMax"
     if args.all:
         grp = ""
+    if args.pattern != None:
+        grp = "| grep %s" % args.pattern
     if args.bake:
         grp = "| grep ' B'"
     if args.pending:
@@ -32,11 +38,14 @@ def main():
     if args.run:
         grp = "| grep RUN_"
     a= subprocess.getoutput ( "slurm q %s" % grp )
+    if args.dry_run:
+        print ( "dry run, but here is what I would do:" )
     print ( "cancelling", end=" " )
     for line in a.split("\n" ):
         jobid = line[:8].strip()
         print ( "%s" % jobid, end=" " )
-        subprocess.getoutput ( "scancel %s" % jobid )
+        if not args.dry_run:
+            subprocess.getoutput ( "scancel %s" % jobid )
     print ( "done." )
 
 main()
